api/pyproject.toml
[tool.poetry]
name = "monayy"
version = "0.1.0"
description = ""
authors = ["Sam Burba <github@samburba.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
dateparser = "^1.1.8"
pytest = "^7.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
aiosqlite = "^0.19.0"
sqlalchemy-utils = "^0.41.1"
pytest-asyncio = "^0.21.1"
black = "^23.7.0"
testing-postgresql = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
